Renamed enum file to statistic-type.enum.ts from datatypes-statistic.enum.ts
