+++ src/app/fyle/my-view-advance/my-view-advance.page.ts
@@ -4,7 +4,7 @@ import { from, Observable } from 'rxjs';
 import { finalize, shareReplay, switchMap } from 'rxjs/operators';
 import { AdvanceService } from 'src/app/core/services/advance.service';
 import { LoaderService } from 'src/app/core/services/loader.service';
-import { StatisticTypes } from 'src/app/shared/components/fy-statistic/datatypes-statistic.enum';
+import { StatisticTypes } from 'src/app/shared/components/fy-statistic/statistic-type.enum';
 import { OfflineService } from 'src/app/core/services/offline.service';
 
 @Component({
