+++ src/app/shared/components/fy-statistic/statistic-type.enum.ts
